Completed forward cube
A cone that will move locally forward and backward with W/S and turn
with A/D and pivot with E/Q. Keep in mind that the cone is symmetrical
and so you cannot tell if it is facing towards or away from you; this
will cause the A/D buttons to react differently. To Do: Replace cone
with a 3D model that has a discernable front, back, and tip.